Added authentication to user routes

+++ trucking-app-flask/routes/v1/user.py
@@ -1,6 +1,7 @@
 from flask import jsonify, request, Blueprint, g
 from sqlalchemy.exc import IntegrityError
 from controllers import UserController
+from flask_login import login_required, current_user
 from models import User
 
 user = Blueprint('user', __name__)
@@ -9,7 +10,11 @@ user = Blueprint('user', __name__)
 
 # GET operation (get user by ID)
 @user.route('/<int:user_id>', methods=['GET'])
+@login_required
 def get_user(user_id):
+    if current_user.id != user_id:
+        return jsonify({"error": "You are not authorized to view this user"}), 403
+
     session = g.session
     return UserController.get_user(session=session, user_id=user_id)
 
@@ -23,7 +28,10 @@ def create_user():
 
 # PUT operation (update user by ID)
 @user.route('/<int:user_id>', methods=['PUT'])
+@login_required
 def update_user(user_id):
+    if current_user.id != user_id:
+        return jsonify({"error": "You are not authorized to update this user"}), 403
     session = g.session
     return UserController.update_user(session=session, request=request, user_id=user_id)
 
@@ -31,7 +39,9 @@ def update_user(user_id):
 
 
 @user.route('/<int:user_id>', methods=['DELETE'])
+@login_required
 def delete_user(user_id):
+    if current_user.id != user_id:
+        return jsonify({"error": "You are not authorized to delete this user"}), 403
     session = g.session
     return UserController.delete_user(session=session, user_id=user_id)
-
